requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-admin-persian-fonts==0.2
django-braces==1.15.0
django-clearcache==1.2.1
django-crispy-forms==1.14.0
django-jazzmin==2.6.0
django-model-utils==4.3.1
django-phonenumber-field==7.0.2
django-taggit==3.1.0
django-tinymce==3.5.0
django-widget-tweaks==1.4.12
idna==3.4
Markdown==3.4.1
Pillow==9.4.0
psycopg2-binary==2.9.5
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
suds==1.1.2
swapper==1.3.0
urllib3==1.26.14
